flower/TABULAR/TABNET/pyproject.toml
[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "quickstart-tabnet"
version = "0.1.0"
description = "Tabnet Federated Learning Quickstart with Flower"
authors = ["The Flower Authors <hello@flower.ai>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
flwr = ">=1.0,<2.0"
tensorflow-cpu = { version = ">=2.9.1,<2.11.1 || >2.11.1", markers = "platform_machine == \"x86_64\"" }
tensorflow-macos = { version = ">=2.9.1,<2.11.1 || >2.11.1", markers = "sys_platform == \"darwin\" and platform_machine == \"arm64\"" }
tensorflow_datasets = "4.9.2"
tabnet = "0.1.6"
flwr-datasets = { extras = ["vision"], version = ">=0.0.2,<1.0.0" }
